Add tests and reuse code among the two test files
