+++ packages/app-store/googlecalendar/lib/__mocks__/features.repository.ts
@@ -1,0 +1,9 @@
+import { vi } from "vitest";
+
+const featuresRepositoryModuleMock = {
+  FeaturesRepository: vi.fn().mockImplementation(() => ({
+    checkIfFeatureIsEnabledGlobally: vi.fn().mockResolvedValue(true),
+  })),
+};
+
+vi.mock("@calcom/features/flags/features.repository", () => featuresRepositoryModuleMock);
