+++ packages/app-store/tests/__mocks__/OAuthManager.ts
@@ -51,5 +51,4 @@ const defaultMockOAuthManager = vi.fn().mockImplementation(() => {
     }),
   };
 });
-
 export { oAuthManagerMock, defaultMockOAuthManager, setFullMockOAuthManagerRequest };
